se agregaron nuevas rutas, correccion de axios

+++ src/lib/axios.js
@@ -1,7 +1,8 @@
+
 import axios from "axios";
 const host = import.meta.env.VITE_HOST_BACKEND
 const port = import.meta.env.VITE_PORT_BACKEND
-const token = localStorage.getItem('token')
+
 
 const api = axios.create({
     baseURL: `http://${host}:${port}/api/v1`    
@@ -23,16 +24,18 @@ api.interceptors.request.use(
 
 // Valida la respuesta recibida para determinar si el token es invalido o ha expirado
 api.interceptors.response.use(
-(response) => response,
-(error) => {
+  (response) => response,
+  (error) => {
     if (error.response && error.response.status === 401) {
-    // Redirigir al login si el token es inválido o ha expirado
-    console.error("Token inválido o expirado");
-    localStorage.clear()
-    router.push({name: 'NoAutenticado'})
+      // Verifica que NO estés ya en login
+      if (window.location.pathname !== '/login') {
+        console.error("Token inválido o expirado");
+        localStorage.clear();
+        window.location.href = "/login";
+      }
     }
     return Promise.reject(error);
-}
+  }
 );
 
 export default api;
